ms-rrhh/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
pydantic
pydantic[email]
alembic
prometheus-client
python-dotenv
structlog
httpx
bcrypt
python-jose[cryptography]
python-multipart
